# artifacts/api-server/src/routes/dashboard.ts
import { Router } from "express";
import { db } from "@workspace/db";
import {
  contactsTable,
  dealsTable,
  ticketsTable,
  leadsTable,
  projectsTable,
  activityTable,
  invoicesTable,
} from "@workspace/db";
import { count, sum, eq, desc } from "drizzle-orm";

const router = Router();

router.get("/dashboard/stats", async (req, res) => {
  try {
    const [
      [{ totalContacts }],
      [{ totalDeals }],
      [{ openTickets }],
      [{ totalLeads }],
      [{ activeProjects }],
      wonDeals,
      allDeals,
      revenueResult,
    ] = await Promise.all([
      db.select({ totalContacts: count() }).from(contactsTable),
      db.select({ totalDeals: count() }).from(dealsTable),
      db
        .select({ openTickets: count() })
        .from(ticketsTable)
        .where(eq(ticketsTable.status, "open")),
      db.select({ totalLeads: count() }).from(leadsTable),
      db
        .select({ activeProjects: count() })
        .from(projectsTable)
        .where(eq(projectsTable.status, "active")),
      db
        .select({ c: count() })
        .from(dealsTable)
        .where(eq(dealsTable.stage, "closed_won")),
      db.select({ c: count() }).from(dealsTable),
      db
        .select({ total: sum(invoicesTable.amount) })
        .from(invoicesTable)
        .where(eq(invoicesTable.status, "paid")),
    ]);

    const winRate =
      allDeals[0].c > 0
        ? Math.round((wonDeals[0].c / allDeals[0].c) * 100)
        : 0;
    const totalRevenue = Number(revenueResult[0].total ?? 0);

    res.json({
      totalRevenue,
      totalContacts,
      totalDeals,
      openTickets,
      totalLeads,
      winRate,
      monthlyGrowth: 12.4,
      activeProjects,
    });
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Failed to fetch stats" });
  }
});

router.get("/dashboard/activity", async (req, res) => {
  try {
    const limit = Number(req.query.limit ?? 20);
    const items = await db
      .select()
      .from(activityTable)
      .orderBy(desc(activityTable.createdAt))
      .limit(limit);
    res.json(
      items.map((i) => ({
        id: i.id,
        type: i.type,
        title: i.title,
        description: i.description,
        createdAt: i.createdAt.toISOString(),
      }))
    );
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Failed to fetch activity" });
  }
});

router.get("/dashboard/revenue-chart", async (req, res) => {
  try {
    const months = Number(req.query.months ?? 6);
    const now = new Date();
    const data = [];
    for (let i = months - 1; i >= 0; i--) {
      const d = new Date(now.getFullYear(), now.getMonth() - i, 1);
      const label = d.toLocaleString("default", {
        month: "short",
        year: "2-digit",
      });
      // Generate plausible demo revenue figures seeded from month index
      const base = 45000 + (months - i) * 8000;
      const variance = Math.sin(i * 1.3) * 12000;
      data.push({
        month: label,
        revenue: Math.round(base + variance),
        deals: 8 + Math.floor(Math.random() * 12),
      });
    }
    res.json(data);
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Failed to fetch revenue chart" });
  }
});

export default router;