# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CINTEXA NEXUS AI Business Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Recent cross-module activity feed
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/revenue-chart:
    get:
      operationId: getRevenueChart
      tags: [dashboard]
      summary: Monthly revenue chart data
      parameters:
        - in: query
          name: months
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Revenue chart data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDataPoint"

  # ── Contacts ───────────────────────────────────────────────────────────
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Companies ──────────────────────────────────────────────────────────
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List all deals
      parameters:
        - in: query
          name: stage
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    patch:
      operationId: updateDeal
      tags: [deals]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Projects ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ActivityItem:
      type: object
      required: [id, type, title, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [contact, deal, ticket, campaign, invoice, task, lead]
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
    RevenueDataPoint:
      type: object
      required: [month, revenue, deals]
      properties:
        month:
          type: string
        revenue:
          type: number
        deals:
          type: integer
...
    Contact:
      type: object
      required: [id, firstName, lastName, email, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, prospect]
        score:
          type: ["integer", "null"]
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        jobTitle:
          type: string
        status:
          type: string
          enum: [active, inactive, prospect]
    ContactUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        jobTitle:
          type: string
        status:
          type: string
          enum: [active, inactive, prospect]
        score:
          type: integer

    # ── Company ──────────────────────────────────────────────────────────
...
    Deal:
      type: object
      required: [id, title, value, stage, probability, contactName, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        value:
          type: number
        stage:
          type: string
          enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
        probability:
          type: integer
        contactName:
          type: string
        companyName:
          type: ["string", "null"]
        expectedCloseDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    DealInput:
      type: object
      required: [title, value, stage, contactName]
      properties:
        title:
          type: string
        value:
          type: number
        stage:
          type: string
          enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
        probability:
          type: integer
        contactName:
          type: string
        companyName:
          type: string
        expectedCloseDate:
          type: string
        notes:
          type: string
    DealUpdate:
      type: object
      properties:
        title:
          type: string
        value:
          type: number
        stage:
          type: string
          enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
        probability:
          type: integer
        contactName:
          type: string
        companyName:
          type: string
        expectedCloseDate:
          type: string
        notes:
          type: string
...
    Invoice:
      type: object
      required: [id, number, clientName, amount, status, dueDate, createdAt]
      properties:
        id:
          type: integer
        number:
          type: string
        clientName:
          type: string
        amount:
          type: number
        tax:
          type: ["number", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: string
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceInput:
      type: object
      required: [clientName, amount, dueDate]
      properties:
        clientName:
          type: string
        amount:
          type: number
        tax:
          type: number
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        clientName:
          type: string
        amount:
          type: number
        tax:
          type: number
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: string
        paidAt:
          type: string

    # ── Project ──────────────────────────────────────────────────────────
...